mofulunches-api/api_gateway/requirements.txt
async-timeout==5.0.1
bidict==0.23.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dbus-python==1.3.2
dnspython==2.7.0
eventlet==0.38.0
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.4.1
future==1.0.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
optimus-manager==1.5
psutil==6.1.0
PyJWT==2.10.0
pymongo==4.10.1
pyserial==3.5
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
serial==0.0.97
simple-websocket==1.1.0
urllib3==2.2.3
Werkzeug==3.1.3
wsproto==1.2.0
